Done with search

+++ app/src/main/java/com/example/beerdiary/data/Beer.kt
@@ -13,5 +13,5 @@ data class Beer(
 )
 
 data class Nutriments(
-    val energy_value : Int
+    val energy_value : Float
 )
